EID-918: Move MetadataAggregatorConfiguration into root package
Move class to help match drop wizard conventions.

Authors: @kgarwood, @felixmarcusmillne
