+++ src/main/java/uk/gov/ida/metadataaggregator/MetadataAggregatorModule.java
@@ -8,7 +8,6 @@ import com.google.inject.Provides;
 import com.google.inject.name.Named;
 import uk.gov.ida.metadataaggregator.core.S3BucketMetadataStore;
 import uk.gov.ida.metadataaggregator.exceptions.ConfigSourceException;
-import uk.gov.ida.metadataaggregator.configuration.MetadataAggregatorConfiguration;
 import uk.gov.ida.metadataaggregator.configuration.MetadataSourceConfiguration;
 import uk.gov.ida.metadataaggregator.configuration.MetadataSourceConfigurationLoader;
 import uk.gov.ida.saml.metadata.EidasTrustAnchorResolver;
